Add serial to id and query to fetch all blog posts.

+++ resources/migrations/init.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXISTS blog_posts (id BIGINT);
+CREATE TABLE IF NOT EXISTS blog_posts (id SERIAL);
